requirements.txt
absl-py==2.1.0
alabaster==0.7.16
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
babel==2.17.0
Box2D==2.3.10
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==3.1.0
-e git+https://github.com/codpy2020/codpy.git@e4e433193ef839fac1f03ab7243a58065de9ea31#egg=codpy
codpydll==0.1.10
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.13
decorator==5.1.1
docutils==0.21.2
etils==1.5.2
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
flatbuffers==25.2.10
fonttools==4.55.3
fsspec==2025.3.0
gast==0.4.0
glfw==2.8.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.69.0
gymnasium==1.0.0
h5py==3.13.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
intel-openmp==2022.2.1
ipykernel==6.29.5
ipython==8.18.1
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kagglehub==0.3.12
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.6.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2022.2.1
ml_dtypes==0.5.1
moviepy==2.1.1
mujoco==3.3.0
namex==0.0.8
narwhals==1.41.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==1.5.0
parso==0.8.4
patsy==1.0.1
pillow==10.4.0
platformdirs==4.3.6
plotly==6.1.1
pockets==0.9.1
proglog==0.1.10
prompt_toolkit==3.0.50
protobuf==3.19.6
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pygame==2.6.1
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==309
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9
scikit-learn==1.6.0
scipy==1.9.1
seaborn==0.13.2
six==1.17.0
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-autodoc-typehints==2.3.0
sphinx-gallery==0.19.0
sphinx-math-dollar==1.2.1
sphinx-rtd-theme==3.0.2
sphinx-togglebutton==0.3.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
statsmodels==0.14.4
tbb==2021.13.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
threadpoolctl==3.5.0
tomli==2.2.1
torch==1.13.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xarray==2022.9.0
xgboost==2.1.4
zipp==3.21.0
